lyria_app/requirements.txt
fastapi
uvicorn
google-genai
websockets
python-multipart
